Add k8s resource deletion and move get model-package/model-package-group to shared file
